detector/requirements.txt
pyyaml>=6.0
psutil>=5.9
